requirements.txt
aniso8601==8.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
environs==8.0.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.6.1
pymongo==3.11.0
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
urllib3==1.25.9
Werkzeug==1.0.1
